+++ tests/test_airflow.py
@@ -5,8 +5,14 @@ from kedro_airflow_k8s.airflow import AirflowClient, MissingDAGException
 
 
 class TestAirflow:
+    @pytest.fixture(scope="class")
+    def client(self):
+        return AirflowClient(
+            "https://test.airflow.com", max_retries=0, retry_interval=0
+        )
+
     @responses.activate
-    def test_get_dag(self):
+    def test_get_dag(self, client):
         response_data = {
             "dag_id": "test_id",
             "tags": [{"name": "commit_sha:123456"}],
@@ -15,9 +21,6 @@ class TestAirflow:
             responses.GET,
             "https://test.airflow.com/api/v1/dags/test_id",
             json=response_data,
-        )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0
         )
 
         dag = client.get_dag("test_id")
@@ -26,7 +29,7 @@ class TestAirflow:
         assert dag.tags == [{"name": "commit_sha:123456"}]
 
     @responses.activate
-    def test_wait_for_dag(self):
+    def test_wait_for_dag(self, client):
         response_data = {
             "dag_id": "test_id",
             "tags": [{"name": "commit_sha:123456"}],
@@ -36,9 +39,6 @@ class TestAirflow:
             "https://test.airflow.com/api/v1/dags/test_id",
             json=response_data,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0
-        )
 
         dag = client.wait_for_dag("test_id", "commit_sha:123456")
 
@@ -46,7 +46,7 @@ class TestAirflow:
         assert dag.tags == [{"name": "commit_sha:123456"}]
 
     @responses.activate
-    def test_wait_for_dag_timeout(self):
+    def test_wait_for_dag_timeout(self, client):
         response_data = {
             "dag_id": "test_id",
             "tags": [{"name": "commit_sha:654321"}],
@@ -56,15 +56,12 @@ class TestAirflow:
             "https://test.airflow.com/api/v1/dags/test_id",
             json=response_data,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0, retry_interval=0
-        )
 
         with pytest.raises(MissingDAGException):
             client.wait_for_dag("test_id", "commit_sha:123456")
 
     @responses.activate
-    def test_trigger_dag_run(self):
+    def test_trigger_dag_run(self, client):
         response_data = {
             "dag_run_id": "test_run_id",
         }
@@ -73,15 +70,12 @@ class TestAirflow:
             "https://test.airflow.com/api/v1/dags/test_id/dagRuns",
             json=response_data,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0
-        )
 
         dag_run_id = client.trigger_dag_run("test_id")
         assert dag_run_id == "test_run_id"
 
     @responses.activate
-    def test_wait_for_dag_run_completion(self):
+    def test_wait_for_dag_run_completion(self, client):
         response_data_dag_run = {"state": "success"}
         response_data_task_instance = {
             "task_instances": [{"state": "success"}]
@@ -97,9 +91,6 @@ class TestAirflow:
             "taskInstances",
             json=response_data_task_instance,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0
-        )
 
         status = client.wait_for_dag_run_completion(
             "test_id", "test_dag_run_id", 1
@@ -108,7 +99,7 @@ class TestAirflow:
         assert status == "success"
 
     @responses.activate
-    def test_wait_for_dag_run_retry(self):
+    def test_wait_for_dag_run_retry(self, client):
         running_response, success_response = {"state": "running"}, {
             "state": "success"
         }
@@ -131,9 +122,6 @@ class TestAirflow:
             "taskInstances",
             json=response_data_task_instance,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0, retry_interval=0
-        )
 
         status = client.wait_for_dag_run_completion(
             "test_id", "test_dag_run_id", 1
@@ -142,14 +130,11 @@ class TestAirflow:
         assert status == "success"
 
     @responses.activate
-    def test_wait_for_dag_run_unknown(self):
+    def test_wait_for_dag_run_unknown(self, client):
         responses.add(
             responses.GET,
             "https://test.airflow.com/api/v1/dags/test_id/dagRuns/test_dag_run_id",
             status=403,
-        )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0, retry_interval=0
         )
 
         status = client.wait_for_dag_run_completion(
@@ -159,7 +144,7 @@ class TestAirflow:
         assert status == "unknown"
 
     @responses.activate
-    def test_wait_for_dag_run_task_instance_failed(self):
+    def test_wait_for_dag_run_task_instance_failed(self, client):
         response_data_dag_run = {"state": "success"}
         response_data_task_instance = {
             "task_instances": [
@@ -178,11 +163,52 @@ class TestAirflow:
             "taskInstances",
             json=response_data_task_instance,
         )
-        client = AirflowClient(
-            "https://test.airflow.com/api/v1", max_retries=0, retry_interval=0
-        )
+
         status = client.wait_for_dag_run_completion(
             "test_id", "test_dag_run_id", 1
         )
 
         assert status == "failed"
+
+    @responses.activate
+    def test_list_dags(self, client):
+        list_of_dags = {
+            "dags": [
+                {
+                    "dag_id": "match",
+                    "tags": [
+                        {"name": "generated_with_kedro_airflow_k8s:0.1.1"},
+                        {"name": "experiment_name:test_experiment"},
+                    ],
+                },
+                {
+                    "dag_id": "not_generated_by_plugin",
+                    "tags": [{"name": "experiment_name:test_experiment"}],
+                },
+                {
+                    "dag_id": "without_experiment_name",
+                    "tags": [
+                        {"name": "generated_with_kedro_airflow_k8s:0.1.1"}
+                    ],
+                },
+            ]
+        }
+
+        responses.add(
+            responses.GET,
+            "https://test.airflow.com/api/v1/dags?limit=1000",
+            json=list_of_dags,
+        )
+
+        dags = client.list_dags("generated_with_kedro_airflow_k8s")
+
+        assert len(dags) == 2
+        assert dags[0].dag_id == "match"
+        assert dags[0].tags == [
+            {"name": "generated_with_kedro_airflow_k8s:0.1.1"},
+            {"name": "experiment_name:test_experiment"},
+        ]
+        assert dags[1].dag_id == "without_experiment_name"
+        assert dags[1].tags == [
+            {"name": "generated_with_kedro_airflow_k8s:0.1.1"}
+        ]