+++ tests/test_cli.py
@@ -1,12 +1,24 @@
+import os
+import webbrowser
 from pathlib import Path
 from tempfile import TemporaryDirectory
 from unittest.mock import MagicMock, Mock, patch
 
+import click
 import pytest
 from click.testing import CliRunner
 
 from kedro_airflow_k8s import airflow
-from kedro_airflow_k8s.cli import compile, run_once, schedule, upload_pipeline
+from kedro_airflow_k8s.airflow import DAGModel
+from kedro_airflow_k8s.cli import (
+    compile,
+    init,
+    list_pipelines,
+    run_once,
+    schedule,
+    ui,
+    upload_pipeline,
+)
 from kedro_airflow_k8s.config import PluginConfig
 from kedro_airflow_k8s.context_helper import ContextHelper
 
@@ -44,19 +56,24 @@ class TestPluginCLI:
         context_helper.project_name = "kedro_airflow_k8s"
         context_helper.config = PluginConfig(
             {
-                "namespace": "test_ns",
-                "image": "test/image:latest",
-                "accessMode": "ReadWriteMany",
-                "requestStorage": "3Gi",
+                "host": "airflow.url.com",
+                "run_config": {
+                    "image": "test/image:latest",
+                    "namespace": "test_ns",
+                    "experiment_name": "kedro_airflow_k8s",
+                    "cron_expression": None,
+                    "volume": {
+                        "access_modes": ["ReadWriteMany"],
+                        "size": "3Gi",
+                        "storageclass": "with-encryption",
+                    },
+                },
             }
         )
         context_helper.mlflow_config = {
             "mlflow_tracking_uri": "mlflow.url.com"
         }
         context_helper.session.store["git"].commit_sha = "abcdef"
-        context_helper.airflow_config = {
-            "airflow_rest_api_uri": "airflow.url.com/api/v1"
-        }
         return context_helper
 
     def test_compile(self, context_helper):
@@ -71,7 +88,7 @@ class TestPluginCLI:
         assert Path("dags/kedro_airflow_k8s.py").exists()
 
         dag_content = Path("dags/kedro_airflow_k8s.py").read_text()
-        assert 'EXPERIMENT_NAME = "kedro_airflow_k8s"' in dag_content
+        assert 'EXPERIMENT_NAME = "kedro-airflow-k8s"' in dag_content
         assert "namespace='test_ns'" in dag_content
         assert "image: image:override" in dag_content
         assert "mlflow_url='mlflow.url.com'" in dag_content
@@ -79,6 +96,7 @@ class TestPluginCLI:
         assert "access_modes=['ReadWriteMany']" in dag_content
         assert "'storage':'3Gi'" in dag_content
         assert "schedule_interval=None" in dag_content
+        assert "storage_class_name='with-encryption'" in dag_content
 
     def test_upload_pipeline(self, context_helper):
         config = dict(context_helper=context_helper)
@@ -229,3 +247,128 @@ class TestPluginCLI:
             Path(output_directory.name) / "kedro_airflow_k8s.py"
         ).read_text()
         assert len(dag_content) > 0
+
+    def test_list_pipelines(self, context_helper):
+        config = dict(context_helper=context_helper)
+
+        runner = CliRunner()
+
+        click.echo = Mock()
+        with patch.object(airflow.AirflowClient, "list_dags") as list_dags:
+            list_dags.return_value = [
+                DAGModel(
+                    dag_id="match0",
+                    tags=[
+                        {"name": "generated_with_kedro_airflow_k8s:0.1.2"},
+                        {"name": "experiment_name:zxw_experiment"},
+                    ],
+                ),
+                DAGModel(
+                    dag_id="match1",
+                    tags=[
+                        {"name": "generated_with_kedro_airflow_k8s:0.1.1"},
+                        {"name": "experiment_name:test_experiment"},
+                    ],
+                ),
+                DAGModel(
+                    dag_id="match2",
+                    tags=[
+                        {"name": "generated_with_kedro_airflow_k8s:0.1.2"},
+                        {"name": "experiment_name:test_experiment"},
+                    ],
+                ),
+            ]
+            result = runner.invoke(
+                list_pipelines,
+                [],
+                obj=config,
+            )
+
+        assert result.exit_code == 0
+        click.echo.assert_called_once()
+        tabulate_result = click.echo.call_args_list[0][0][0]
+        assert "match0" in tabulate_result
+        assert "match1" in tabulate_result
+        assert "match2" in tabulate_result
+        assert "test_experiment" in tabulate_result
+        assert "zxw_experiment" in tabulate_result
+
+    def test_ui(self, context_helper):
+        config = dict(context_helper=context_helper)
+
+        runner = CliRunner()
+
+        webbrowser.open_new_tab = Mock()
+
+        result = runner.invoke(
+            ui,
+            [],
+            obj=config,
+        )
+
+        assert result.exit_code == 0
+        webbrowser.open_new_tab.assert_called_once_with("airflow.url.com")
+
+    def test_ui_with_dag_view(self, context_helper):
+        config = dict(context_helper=context_helper)
+
+        runner = CliRunner()
+
+        webbrowser.open_new_tab = Mock()
+
+        result = runner.invoke(
+            ui,
+            ["--dag-name", "test-dag"],
+            obj=config,
+        )
+
+        assert result.exit_code == 0
+        webbrowser.open_new_tab.assert_called_once_with(
+            "airflow.url.com/tree?dag_id=test-dag"
+        )
+
+    def test_init(self, context_helper):
+        context_helper.context.project_path = Path("test-name")
+        config = dict(context_helper=context_helper)
+
+        runner = CliRunner()
+
+        cwd = os.getcwd()
+        try:
+            with TemporaryDirectory() as td:
+                os.chdir(td)
+                result = runner.invoke(
+                    init,
+                    [
+                        "--with-github-actions",
+                        "--output=gs://dag.bucket",
+                        "https://test.apache.airflow.com",
+                    ],
+                    obj=config,
+                )
+
+                assert result.exit_code == 0
+
+                assert (Path(td) / "conf/base/airflow-k8s.yaml").exists()
+                assert (
+                    "{{"
+                    not in (
+                        Path(td) / "conf/base/airflow-k8s.yaml"
+                    ).read_text()
+                )
+                assert (
+                    Path(td) / ".github/workflows/on-merge-to-master.yml"
+                ).exists()
+                assert (
+                    "PROJECT_NAME: test-name"
+                    in (
+                        Path(td) / ".github/workflows/on-merge-to-master.yml"
+                    ).read_text()
+                )
+                assert (Path(td) / ".github/workflows/on-push.yml").exists()
+                assert (
+                    "PROJECT_NAME: test-name"
+                    in (Path(td) / ".github/workflows/on-push.yml").read_text()
+                )
+        finally:
+            os.chdir(cwd)